Moving dag optimizations to their own file
